Fix rake never running the test/test_*.rb suite
test_all invokes Rake::Task['test'], but the Rake::TestTask that defined it
was commented out. Rake therefore synthesized a file task for the existing
test/ directory, which has no actions, so the invoke was a no-op that
neither ran anything nor complained.

    Rake::Task.task_defined?("test")  # => false
    Rake::Task["test"].class          # => Rake::FileTask
    Rake::Task["test"].actions.size   # => 0

So `bundle exec rake` -- the default task, and what CI runs -- executed only
tests.rb, tests_mimic.rb, tests_mimic_addition.rb and test/json_gem. None of
the test/test_*.rb files ran. Most of them appear in the memcheck list,
which covers one platform and one Ruby; the rest ran nowhere at all:
test_debian, test_generate, test_parser, test_parser_saj, test_realworld,
test_scp, test_various. `bundle exec rake` goes from 569 runs and 2722
assertions to 1312 and 14496.

Each file runs in its own process, the way test_all already runs
test/json_gem. Every one of them passes alone, but loading them all into one
process does not work: they share Oj.default_options, the Oj::Parser.saj and
Oj::Parser.usual singletons and Oj.mimic_JSON, so results depend on
collection order. With a single shared process the suite passes on seeds 1
and 2 and fails with 9 to 13 failures on seeds 42, 1234 and 9999. Isolating
that state is worth doing separately -- it is a change to the tests, not to
how they are collected.

The list of files is now in one place, so a new test/test_*.rb is picked up
by `rake test` and `rake test:valgrind` without being added by hand
anywhere. The memcheck list is that same set minus two files, since
ruby_memcheck loads everything it is given into one process:

  * test_generate.rb calls Oj.generate, which calls Oj::Rails.mimic_JSON
    implicitly, after which a parse error is a JSON::ParserError rather than
    an Oj::ParseError and the tests in test_max_integer_digits.rb that name
    the class fail. Nothing undoes that, which is why the file belongs
    alongside test/isolated/.
  * test_scp_fork.rb is new, holding the two Oj.sc_parse tests that fork a
    writer for the pipe they read. Valgrind traces the child as well and
    both processes write to the one XML report, leaving it truncated:
    "Premature end of data in tag valgrindoutput". Splitting them out keeps
    the other 21 tests of test_scp.rb under the lane, which also settles
    the interference with test_various that kept the whole file out of it.

That takes the memcheck lane from 18 files to 24, at 690 runs and 11,356
assertions rather than 656 and 11,318, and it is stable across seeds 1, 42
and 1234.

Part of that gain is a name collision: test_saj.rb and test_parser_saj.rb
both defined SajTest, so in a shared process whichever loaded second
discarded 22 of the other's tests, with only a -w warning to say so. The
parser one is now ParserSajTest. Both files also defined an identical
AllSaj, which is now test/saj_handlers.rb, next to the test/scp_handlers.rb
the two scp files share.

Two tests had rotted unnoticed while nothing ran them:

  * test_parser_saj.rb's test_file opened 'saj_test.json', a fixture that
    was never committed, so it could not have passed. Added the fixture the
    assertions describe and made the path relative to __dir__ rather than to
    the working directory.

  * test_debian.rb's test_as_json_object_compat_hash_cached expects
    :use_as_json to be honoured in :compat mode. It is not: neither as_json
    nor to_json is called and the object is dumped as its to_s, so Oj.load
    returns a String. :rails and :custom do call as_json. Whether :compat
    should honour it is a question about intended semantics rather than test
    plumbing, so the test is skipped with that evidence recorded instead of
    having its expectation edited.
