+++ test/saj_test.json
@@ -1,0 +1,1 @@
+[true,false]
