+++ test/test_saj.rb
@@ -4,6 +4,7 @@
 $LOAD_PATH << __dir__
 
 require 'helper'
+require 'saj_handlers'
 
 $json = %{{
   "array": [
@@ -19,41 +20,6 @@ $json = %{{
   ],
   "boolean" : true
 }}
-
-class AllSaj < Oj::Saj
-  attr_accessor :calls
-
-  def initialize
-    @calls = []
-
-    super
-  end
-
-  def hash_start(key)
-    @calls << [:hash_start, key]
-  end
-
-  def hash_end(key)
-    @calls << [:hash_end, key]
-  end
-
-  def array_start(key)
-    @calls << [:array_start, key]
-  end
-
-  def array_end(key)
-    @calls << [:array_end, key]
-  end
-
-  def add_value(value, key)
-    @calls << [:add_value, value, key]
-  end
-
-  def error(message, line, column)
-    @calls << [:error, message, line, column]
-  end
-
-end # AllSaj
 
 class SajTest < Minitest::Test
